build: use tinacms-scripts to build package
