+++ packages/gatsby-tinacms-json/.gitignore
@@ -1,2 +1,2 @@
-*.js
+index.js
 *.d.ts
